Re-estructure a lot of code.
