+++ commands/general/afk.js
@@ -15,10 +15,10 @@ module.exports = {
 	 * @param {Array} args
 	 * @param {String} prefix
 	 */
-	execute(client, message, args, prefix) {
+	async execute(client, message, args, prefix) {
 		let actual_time = new Date().getTime();
         let reason = args.slice(0).join(" ");
 		client.user_data.prepare("INSERT OR REPLACE INTO afk (user_id, name, reason, time) VALUES (?, ?, ?, ?);").run(message.author.id, message.author.tag, reason, actual_time);
-		return message.channel.send(client.functions.getTranslation(client, message.author, message.guild, "commands_afk", "set", [message.author.tag]) + "\n\n" + reason);
+		return message.inlineReply(client.functions.getTranslation(client, message.author, message.guild, "commands_afk", "set", [message.author.tag]) + "\n\n" + reason);
 	},
 };
