+++ commands/information/member.js
@@ -40,7 +40,7 @@ module.exports = {
 		
 		let mentioned_member = message.mentions.members.first();
         let member = mentioned_member || get_member;
-		if (!member) { if (args[0]) { member = await client.fetchers.getGuildMember(client, message.guild, args[0]); } else { member = message.member; } }
+		if (!member) { if (args[0]) { member = await message.guild.members.fetch(args[0]); } else { member = message.member; } }
 		if (!member) {
 			let embed = new Discord.MessageEmbed();
 			embed.setTitle(client.functions.getTranslation(client, message.author, message.guild, "command.member.failure.title"));
