+++ functions/events/guilds_bot_welcome.js
@@ -2,14 +2,14 @@ const Discord = require("discord.js");
 async function first_time_welcome(client, guild) {
 	console.log("Guild joined " + "'" + guild.name + "'" + "." + " (" + guild.id + ")");
 	
-	let channel_target = undefined;
+	let channel_target;
 	
 	
 	let welcome_description = [
-		"Gracias por añadirme a su servidor y es un placer estar aqui.",
+		"Gracias por aÃ±adirme a su servidor y es un placer estar aqui.",
 		"Soy un bot con caracteristicas posiblemente interesantes y detalladas.",
 		"Para comenzar a saber como funciono escriba `=help`.",
-		"Para saber la lista de cambios use `=changelog`. (Solo disponible en español).",
+		"Para saber la lista de cambios use `" + client.config.default.prefix + "changelog`. (Solo disponible en espaÃ±ol).",
 	];
 	
 	let bot_owner = client.users.cache.get(client.config.owner);
