--- functions/general/crypto.js
@@ -1,20 +1,0 @@
-const crypto = require("crypto");
-const algorithm = "aes-256-ctr";
-const secretKey = "vOVH6sdmpNWjRRIqCc7rdxs01lwHzfr3";
-const iv = crypto.randomBytes(256);
-const encrypt = (text) => {
-    const cipher = crypto.createCipheriv(algorithm, secretKey, iv);
-    const encrypted = Buffer.concat([cipher.update(text), cipher.final()]);
-    return {iv: iv.toString("hex"), content: encrypted.toString("hex")};
-};
-
-const decrypt = (hash) => {
-    const decipher = crypto.createDecipheriv(algorithm, secretKey, Buffer.from(hash.iv, "hex"));
-    const decrpyted = Buffer.concat([decipher.update(Buffer.from(hash.content, "hex")), decipher.final()]);
-    return decrpyted.toString();
-};
-
-module.exports = {
-	encrypt,
-    decrypt
-};
