+++ functions/main/time_functions.js
@@ -1,6 +1,3 @@
-function getRandomInt(max) { return Math.floor(Math.random() * Math.floor(max)); }
-function getRandomRangeInt(min, max) { return Math.floor(Math.random() * ((max - min) + 1)) + min; }
-
 function generateDateString(client, author, guild, get_date) {
 	if (typeof get_date !== "object") { return; }
 	
@@ -87,161 +84,8 @@ function ISODateToJSDate(isodate_string) {
     return new Date(parts[0], parts[1] - 1, parts[2], parts[3], parts[4], parts[5]);
 }
 
-function numberOrdinal(number) {
-	if (number <= 0) { return number; } // Zero or negatives doesn't supported
-	if (((number % 100) >= 11) && ((number % 100) <= 13)) { return number + "th"; } // More or equals that 11 and less or equals that 13
-	switch (number % 10) // Return depending number
-	{
-		case 1: { return number + "st"; } // One
-		case 2: { return number + "nd"; } // Two
-		case 3: { return number + "rd"; } // Three
-		default: { return number + "th"; } // Four or more
-	}
-}
-
-function number_formatter(num, digits) {
-	var formats = [
-		{value: 1, symbol: ""}, // None
-		{value: 1E3, symbol: "k"}, // Thousand
-		{value: 1E6, symbol: "M"}, // Million
-		{value: 1E9, symbol: "B"}, // Billion
-		{value: 1E12, symbol: "T"}, // Trillion
-		{value: 1E15, symbol: "QD"}, // Quadrillion
-		{value: 1E18, symbol: "QT"}, // Quintillion
-		{value: 1E21, symbol: "SX"}, // Sextillion
-		{value: 1E24, symbol: "SP"}, // Septillion
-		{value: 1E27, symbol: "OC"}, // Octillion
-		{value: 1E30, symbol: "NO"}, // Nonillion
-		{value: 1E33, symbol: "DE"}, // Decillion
-		{value: 1E36, symbol: "UN"}, // Undecillion
-		{value: 1E39, symbol: "DU"}, // Duodecillion
-		{value: 1E42, symbol: "TR"}, // Tredecillion
-		{value: 1E45, symbol: "QU"}, // Quattuordecillion
-		{value: 1E48, symbol: "QN"}, // Quindecillion
-		{value: 1E51, symbol: "SD"}, // Sexdecillion
-		{value: 1E54, symbol: "SP"}, // Septendecillion
-		{value: 1E57, symbol: "OD"}, // Octodecillion	
-		{value: 1E60, symbol: "ND"}, // Novemdecillion
-		{value: 1E63, symbol: "VT"}, // Vigintillion
-		{value: 1E66, symbol: "CT"} // Centillion
-	];
-	
-	let result = num;
-	let rx = /\.0+$|(\.[0-9]*[1-9])0+$/;
-	for (let i = formats.length - 1; i > 0; i--) {
-		if (num >= formats[i].value) {
-			result = (num / formats[i].value).toFixed(digits).replace(rx, "$1") + formats[i].symbol;
-			break;
-		}
-	}
-	return result;
-}
-
-function resources_monitoring(client) {
-	// CPU usage
-	let previousTime = new Date().getTime();
-	let previousUsage = process.cpuUsage();
-	let lastUsage;
-	setInterval(() => {
-		let currentUsage = process.cpuUsage(previousUsage);
-		previousUsage = process.cpuUsage();
-		
-		let currentTime = new Date().getTime();
-		let timeDelta = (currentTime - previousTime) * 10;
-		let { user, system } = currentUsage;
-		lastUsage = { system: system / timeDelta, total: (system + user) / timeDelta, user: user / timeDelta };
-		previousTime = currentTime;
-		client.statsCPU = lastUsage;
-		if (lastUsage.total > 170) { process.abort(); }
-	}, 1000);
-
-	// RAM usage
-	setInterval(() => { if ((process.memoryUsage().rss / 1000000) > 512) { process.abort(); } }, 1000);
-}
-
-function create_server_data(client, guild) {
-	if (!guild) { return; }
-	
-	let get_server_settings = client.server_data.prepare("SELECT * FROM settings WHERE guild_id = ?;").get(guild.id);
-	if (!get_server_settings) {
-		try {
-			client.server_data.prepare("INSERT INTO settings (guild_id, prefix, language) VALUES (?, ?, ?);").run(guild.id, client.config.default.prefix, client.config.default.language);
-		}
-		catch (error) {
-			console.error("ERROR: " + "Cannot create settings for guild: " + guild.name + "\n", error);
-		}
-		finally {
-			console.log("Created settings for guild: " + guild.name);
-		}
-	}
-	
-	let get_server_features = client.server_data.prepare("SELECT * FROM features WHERE guild_id = ?;").get(guild.id);
-	if (!get_server_features) {
-		try {
-			client.server_data.prepare("INSERT INTO features (guild_id, disabled_functions, disabled_commands) VALUES (?, ?, ?);").run(guild.id, "", "");
-		}
-		catch (error) {
-			console.error("ERROR: " + "Cannot create features for guild: " + guild.name + "\n", error);
-		}
-		finally {
-			console.log("Created features for guild: " + guild.name);
-		}
-	}
-}
-
-function create_user_data(client, user) {
-	if (!user) { return; }
-	if (user.bot) { return; }
-	
-	let get_user_settings = client.user_data.prepare("SELECT * FROM settings WHERE user_id = ?;").get(user.id);
-	if (!get_user_settings) {
-		try {
-			client.user_data.prepare("INSERT INTO settings (user_id, language) VALUES (?, ?)").run(user.id, client.config.default.language);
-		}
-		catch (error) {
-			console.error("ERROR: " + "Cannot create settings for user: " + user.tag + "\n", error);
-		}
-		finally {
-			console.log("Created settings for user: " + user.tag);
-		}
-	}
-}
-
-// Status updater
-function status_update(client) {
-	let status_data = require(process.cwd() + "/configurations/activity.js");
-    client.user.setPresence(status_data[getRandomInt(status_data.length)]);
-}
-
-async function generateRankIcon(client, Canvas, level_number) {
-	let level_table = client.config.exp_shield_table;
-	let rank_front_image;
-	let rank_back_image;
-	let get_backlayer = level_table.find(level_table_index => level_number >= level_table_index.level);
-	if (get_backlayer) {
-		rank_back_image = await Canvas.loadImage(process.cwd() + "/assets/images/ranking/backlayer/rank_back_icon_" + get_backlayer.type + ".png");
-	}
-	
-	if (level_number > -1) {
-		rank_front_image = await Canvas.loadImage(process.cwd() + "/assets/images/ranking/frontlayer/rank_front_icon_" + (level_number % 60) + ".png");
-	}
-	
-	if (level_number >= client.config.exp_level_max) {
-		rank_front_image = await Canvas.loadImage(process.cwd() + "/assets/images/ranking/frontlayer/rank_front_icon_60.png");
-	}
-	return {rank_front_image: rank_front_image, rank_back_image: rank_back_image};
-}
-
 module.exports = {
-	create_server_data: create_server_data,
-	create_user_data: create_user_data,
-	getRandomInt: getRandomInt,
-	number_formatter: number_formatter,
-	resources_monitoring: resources_monitoring,
-	status_update: status_update,
-	generateRankIcon: generateRankIcon,
-	numberOrdinal: numberOrdinal,
-	ISODateToJSDate: ISODateToJSDate,
+    ISODateToJSDate: ISODateToJSDate,
 	generateDateString: generateDateString,
 	generateTimeString: generateTimeString,
 	generateDurationString: generateDurationString
