+++ main.js
@@ -45,8 +45,6 @@ const client = new Discord.Client({presence: {status: "invisible"}, fetchAllMemb
 client.config = require(process.cwd() + "/configurations/client.js");
 client.functions = require(process.cwd() + "/functions/general/core.js");
 client.functions.getTranslation = require(process.cwd() + "/functions/general/languages.js");
-client.crypt = require(process.cwd() + "/functions/general/crypto.js");
-client.fetchers = require(process.cwd() + "/functions/general/fetchers.js");
 client.commands = new Discord.Collection();
 client.cooldowns = new Discord.Collection();
 client.exp_cooldowns = new Discord.Collection();
