+++ web/server_leaderboard.js
@@ -8,17 +8,16 @@ module.exports = {
 		
 		var get_header = await fs.readFileSync(process.cwd() + "/assets/web/html/main_header.html", "utf8");
 		
-		let target_guild = undefined;
+		let target_guild;
 		let target_id = request.params.id;
-		if (target_id) {
-			target_guild = await client.fetchers.getGuild(client, target_id);
-		}
+		if (target_id) { target_guild = await client.guilds.fetch(target_id); }
 		if (!target_guild) {
 			let variable_data = {
 				web_header: get_header,
 				bot_image: client.user.displayAvatarURL({format: "png", dynamic: true, size: 64}),
 				error_message: "Invalid guild ID!"
-			}
+			};
+
 			fs.readFile(process.cwd() + "/assets/web/html/error.html", "utf8", function(err, data) {
 				if (err) { console.error(err); }
 				
@@ -41,7 +40,8 @@ module.exports = {
 				web_header: get_header,
 				bot_image: client.user.displayAvatarURL({format: "png", dynamic: true, size: 64}),
 				error_message: "Specified guild disabled Ranking system!"
-			}
+			};
+
 			fs.readFile(process.cwd() + "/assets/web/html/error.html", "utf8", function(err, data) {
 				if (err) { console.error(err); }
 				
@@ -59,7 +59,7 @@ module.exports = {
 		
 		let members_levels = "<table class=\"scoreboard_table\"><tr><th>Rank</th><th>Username</th><th>Level</th><th>Messages</th></tr>";
 		let levels_database = client.server_data.prepare("SELECT * FROM exp WHERE guild_id = ? ORDER BY score DESC;").all(target_guild.id); //client.server_level.list.all(message.guild.id);
-		let members_get = await client.fetchers.getGuildMembers(client, target_guild);
+		let members_get = await target_guild.members.fetch();
 		for (let level_index = 0; level_index < levels_database.length; level_index++) {
 			let level_element = levels_database[level_index];
 			
@@ -69,7 +69,7 @@ module.exports = {
 			let exp_score_base = client.config.exp_score_base;
 			let score_actual = (level_element.level * level_element.level) * exp_score_base;
 			let score_goal = (next_level * next_level) * exp_score_base;
-			let score_progress = ((level_element.score - score_actual) / (score_goal - score_actual)) * 100
+			let score_progress = ((level_element.score - score_actual) / (score_goal - score_actual)) * 100;
 			
 			let rank = (level_index + 1);
 			let member_find = members_get.find(member => member.user.id === level_element.user_id);
@@ -80,7 +80,7 @@ module.exports = {
 				member_id = member_find.user.id;
 			}
 			else {
-				let user_find = await client.fetchers.getUser(client, level_element.user_id);
+				let user_find = await client.users.fetch(client, level_element.user_id);
 				if (user_find) {
 					member_name = user_find.tag;
 					member_id = user_find.id;
@@ -95,7 +95,7 @@ module.exports = {
 			bot_image: client.user.displayAvatarURL({format: "png", dynamic: true, size: 64}),
 			server_name: target_guild.name,
 			server_leaderboard: members_levels
-		}
+		};
 		
 		fs.readFile(process.cwd() + "/assets/web/html/server_leaderboard.html", "utf8", function(err, data) {
 			if (err) { console.error(err); }
@@ -109,4 +109,4 @@ module.exports = {
 			return response.send(html);
 		});
 	}
-}
+};
